+++ esp-wifi/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `AccessPointInfo::country` to access the Country Code from the Wi-Fi scan results (#3837)
+
 
 ### Changed